+++ services/workflows-service/scripts/workflows/ui-definition/kyb-with-associated-companies/definition/compose-child-associated-company-definition.ts
@@ -57,7 +57,6 @@ export const composeChildAssociatedCompanyDefinition = ({
       apiPlugins: [],
     },
     config: {
-      createCollectionFlowToken: true,
       workflowLevelResolution: true,
     },
     projectId: projectId,