+++ services/workflows-service/src/workflow/workflow.service.unit.test.ts
@@ -102,7 +102,6 @@ describe('WorkflowService', () => {
   let projectScopeService;
   let businessRepo;
   let businessService;
-  let businessReportService;
   let customerService;
   let endUserRepo;
   let entityRepo;
@@ -135,7 +134,6 @@ describe('WorkflowService', () => {
     workflowRuntimeDataRepo = new FakeWorkflowRuntimeDataRepo();
     businessRepo = new FakeBusinessRepo();
     businessService = new FakeBusinessRepo();
-    businessReportService = new FakeBusinessRepo();
     endUserRepo = new FakeEndUserRepo();
     entityRepo = new FakeEntityRepo();
     customerService = new FakeCustomerRepo();
@@ -188,7 +186,6 @@ describe('WorkflowService', () => {
       workflowDefinitionRepo as any,
       workflowRuntimeDataRepo,
       endUserRepo,
-      businessReportService,
       {} as any,
       businessRepo,
       businessService,